Update paapi.py
The current "parse only what's needed" approach leaves the developer completely unable to select additional keys which are available in the underlying data model. By pointing the product to the amz `item`, the developer has the ability to fetch whatever information is needed, which may not be available via the library's accessors.